+++ internship-managment-application/src/components/PastFormInfo.vue
@@ -1,0 +1,9 @@
+<template>
+    <div>
+    </div>
+</template>
+<script>
+export default {
+    name: 'PastFormInfo'
+}
+</script>
