+++ internship-managment-application/src/router/index.js
@@ -76,6 +76,7 @@ const routes = [
     path: '/student-internship-add-form',
     name: 'internship-add-form',
     component: () => import('../views/StudentInternshipAddForm.vue'),
+    redirect: '/student-internship-add-form/past-forms',
     meta: {
       title: 'Internship Add Form'
     },
@@ -89,11 +90,11 @@ const routes = [
         }
       },
       {
-        path: 'start-new-form',
-        name: 'internship-add-start-new-form',
-        redirect: '/student-internship-add-form',
+        path: 'new-form',
+        name: 'internship-add-form-new-form',
+        component: () => import('../views/StudentAddFormNew.vue'),
         meta: {
-          title: 'Start New Form'
+          title: 'Internship Add Form - New Form'
         }
       }
     ]
